backend/config/security.py
"""
Security configuration — JWT tokens, password hashing, auth dependencies.
"""

import os
import logging
from datetime import datetime, timedelta
from typing import Optional
from uuid import UUID

from fastapi import Depends, HTTPException, Request, status
from jose import JWTError, jwt
from passlib.context import CryptContext
from sqlalchemy.orm import Session

from config.database import get_db

logger = logging.getLogger(__name__)

# ── Settings ──
SECRET_KEY = os.getenv("SECRET_KEY", "lexguard-dev-secret-change-in-production")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "1440"))  # 24h default
COOKIE_NAME = "access_token"

# ── Password hashing ──
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


def hash_password(password: str) -> str:
    return pwd_context.hash(password)


def verify_password(plain: str, hashed: str) -> bool:
    return pwd_context.verify(plain, hashed)


# ── JWT ──
def create_access_token(user_id: UUID, expires_delta: Optional[timedelta] = None) -> str:
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
    payload = {"sub": str(user_id), "exp": expire}
    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)


def decode_access_token(token: str) -> Optional[str]:
    """Return user_id string or None on failure."""
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload.get("sub")
    except JWTError:
        return None


# ── FastAPI dependencies ──
def get_current_user(request: Request, db: Session = Depends(get_db)):
    """Require authenticated user — raise 401 if missing/invalid token."""
    from models.db_models import UserDB

    token = request.cookies.get(COOKIE_NAME)
    if not token:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Требуется авторизация",
        )
    user_id_str = decode_access_token(token)
    if user_id_str is None:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Недействительный токен",
        )
    try:
        user = db.get(UserDB, UUID(user_id_str))
    except (ValueError, Exception):
        user = None

    if user is None:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Пользователь не найден",
        )
    return user


def get_optional_user(request: Request, db: Session = Depends(get_db)):
    """Return current user or None — no 401 raised."""
    from models.db_models import UserDB

    token = request.cookies.get(COOKIE_NAME)
    if not token:
        return None
    user_id_str = decode_access_token(token)
    if user_id_str is None:
        return None
    try:
        return db.get(UserDB, UUID(user_id_str))
    except Exception:
        return None
