backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
requests==2.32.3
pydantic==2.8.0
qdrant-client==1.9.1
sentence-transformers==3.0.1
torch==2.3.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
celery[redis]==5.4.0
redis==5.2.1
